requirements.txt
torch
torchvision
numpy
pillow

matplotlib
opencv-python

torchmetrics>=1.2.0
lpips>=0.1.4
